refactor: apply pull-request comments
